Move RSS generator to separate package
Storing RSS data in a plist will mean we can offer a generic data
structure and exporter that can be used with other packages.

It also means it will be possible to cache RSS data.  This will give
us the ability to skip HTML export for files that have not changed
making org-publish-rss more efficient.